Improved flexibility

+++ functions/api/queue.js
@@ -5,12 +5,14 @@ const db = admin.firestore();
 
 const queueApi = express.Router();
 
-// Join 9x9 queue
-queueApi.post("/queue/9/join", checkIfAuthenticated, (req, res) => {
+// Join queue
+queueApi.post("/queue/:size/join", checkIfAuthenticated, (req, res) => {
   (async () => {
     try {
       const uid = req.authId;
       const user = db.collection("status").doc(uid);
+      const size = req.params.size;
+
       let presence = await user.get();
       // Check if user exists
       if (!presence.exists) {
@@ -21,34 +23,50 @@ queueApi.post("/queue/9/join", checkIfAuthenticated, (req, res) => {
         console.log(`Warning: User ${uid} is offline!`);
       }
 
+      if ([9, 13, 19].indexOf(parseInt(size)) === -1 && size !== "seasonal") {
+        throw Error(
+          `Only size 9, 13, 19 or seasonal queues are allowed! Given param: ${size}`
+        );
+      }
+
+      // Check if user has already in a matched game
+      const match = await db.collection("matching").doc(uid).get();
+      if (match.exists) {
+        throw Error(`User ${uid} is still in a matched game!`);
+      }
+
       const userData = (await db.collection("users").doc(uid).get()).data();
 
       await db
-        .collection("queue_9")
-        .doc("/" + uid + "/")
+        .collection("queue_" + size)
+        .doc(uid)
         .create({
           elo: userData.elo,
-          created_at: Date.now(),
+          created_at: admin.firestore.FieldValue.serverTimestamp(),
         })
-        .then(console.log(`Player ${uid} joined 9x9 queue!`));
-      return res.status(200).send({ success: "Joined 9x9 queue!" });
+        .then(console.log(`Player ${uid} joined size ${size} queue!`));
+      return res.status(200).send({ success: `Joined ${size}x${size} queue!` });
     } catch (error) {
       console.log(error);
-      return res.status(500).send({ error: error });
+      return res.status(500).send({ error: error.message });
     }
   })();
 });
 
 // Quit 9x9 queue
-queueApi.delete("/queue/9/remove", checkIfAuthenticated, (req, res) => {
+queueApi.delete("/queue/:size/remove", checkIfAuthenticated, (req, res) => {
   (async () => {
     try {
-      const player = db.collection("queue_9").doc(req.authId);
+      const player = db.collection("queue_" + req.params.size).doc(req.authId);
+      const playerDetails = await player.get();
+      if (!playerDetails.exists) {
+        throw Error("Player did not join this queue!");
+      }
       await player.delete();
-      return res.status(200).send({ success: "Removed from 9x9 queue." });
+      return res.status(200).send({ success: "Removed from queue." });
     } catch (error) {
       console.log(error);
-      return res.status(500).send(error);
+      return res.status(500).send({ error: error.message });
     }
   })();
 });
